pyproject.toml
[tool.poetry]
name = "y"
version = "0.1.0"
description = ""
authors = ["Denys Filichkin <denys.philichkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.3.0"}
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
alembic = "^1.11.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
httpx = "^0.24.1"
cloudinary = "^1.34.0"
qrcode = "^7.4.2"
libgravatar = "^1.0.4"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
python-multipart = "^0.0.6"
redis = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
